requirements.txt
openai~=1.63.2
azure-search-documents==11.5.1
azure-core==1.31.0
python-dotenv~=1.0.1
Quart~=0.19.6
quart-cors==0.7.0
azure-identity==1.17.1
aiohttp==3.10.11
openai-messages-token-helper==0.1.10
PyYAML~=6.0.2
Flask~=3.0.3
python-jose~=3.4.0
tenacity~=8.5.0
uvicorn==0.30.1
pytz==2024.1
beautifulsoup4~=4.12.3
msal~=1.30.0
requests~=2.32.3
llm-guard~=0.3.15
openpyxl~=3.1.5
rouge_score~=0.1.2
pandas~=2.2.3
torch~=2.6.0
nltk~=3.9.1
scikit-learn~=1.5.1
transformers~=4.48.2